1 Basic save game - overrides on new game
First implementation of the save game. Only 1 slot available. No "Load UI". It will save the timer seconds and minutes as well as the player location. Need to add load UI, multiple saves and all of the interactable items to save_bp once they are created.
I will return to this when I am done everything else that is more important.